add publisher
