+++ src/AmqpTools.sln
@@ -11,6 +11,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AmqpShovel", "AmqpShovel\AmqpShovel.csproj", "{34A7F279-4D6F-4135-93EF-9BFE42007E50}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AmqpPublisher", "AmqpPublisher\AmqpPublisher.csproj", "{4B83DC28-E437-4622-97A3-79010183E828}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AmqpCommon", "AmqpCommon\AmqpCommon.csproj", "{8EA22BC0-44F6-44CE-8176-D5C3E3607DA2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +25,14 @@ Global
 		{34A7F279-4D6F-4135-93EF-9BFE42007E50}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{34A7F279-4D6F-4135-93EF-9BFE42007E50}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{34A7F279-4D6F-4135-93EF-9BFE42007E50}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4B83DC28-E437-4622-97A3-79010183E828}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4B83DC28-E437-4622-97A3-79010183E828}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4B83DC28-E437-4622-97A3-79010183E828}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4B83DC28-E437-4622-97A3-79010183E828}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8EA22BC0-44F6-44CE-8176-D5C3E3607DA2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8EA22BC0-44F6-44CE-8176-D5C3E3607DA2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8EA22BC0-44F6-44CE-8176-D5C3E3607DA2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8EA22BC0-44F6-44CE-8176-D5C3E3607DA2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
